CSN-95527: move Git related classes to a storage implementation package. Remove Gene*** references, that are profane now.
