+++ pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.genesis.whitelist</groupId>
+    <groupId>com.gn.whitelist</groupId>
     <artifactId>whitelisting-tool</artifactId>
     <version>1.0.0-SNAPSHOT</version>
 
@@ -221,9 +221,9 @@
                             <configOptions>
                                 <useTags>true</useTags>
                                 <sourceFolder>.</sourceFolder>
-                                <basePackage>com.genesis.whitelist</basePackage>
-                                <apiPackage>com.genesis.whitelist</apiPackage>
-                                <modelPackage>com.genesis.whitelist.model</modelPackage>
+                                <basePackage>com.gn.whitelist</basePackage>
+                                <apiPackage>com.gn.whitelist</apiPackage>
+                                <modelPackage>com.gn.whitelist.model</modelPackage>
                                 <performBeanValidation>true</performBeanValidation>
                                 <openApiNullable>false</openApiNullable>
                                 <documentationProvider>none</documentationProvider>
