--- src/main/java/com/genesis/whitelist/exceptions/OperatorAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package com.genesis.whitelist.exceptions;
-
-public class OperatorAlreadyExistsException extends RuntimeException{
-    public OperatorAlreadyExistsException(String operator){
-        super("Operator [" + operator + "] already exists");
-    }
-}
