--- src/main/java/com/genesis/whitelist/exceptions/OperatorMissingException.java
@@ -1,8 +1,0 @@
-package com.genesis.whitelist.exceptions;
-
-
-public class OperatorMissingException extends RuntimeException{
-    public OperatorMissingException(String operator){
-        super("Operator [" + operator + "] doesn't have a file in the repo");
-    }
-}
