+++ src/main/java/com/gn/whitelist/exceptions/OperatorAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package com.gn.whitelist.exceptions;
+
+public class OperatorAlreadyExistsException extends RuntimeException{
+    public OperatorAlreadyExistsException(String operator){
+        super("Operator [" + operator + "] already exists");
+    }
+}
