+++ src/main/java/com/gn/whitelist/exceptions/OperatorMissingException.java
@@ -1,0 +1,8 @@
+package com.gn.whitelist.exceptions;
+
+
+public class OperatorMissingException extends RuntimeException{
+    public OperatorMissingException(String operator){
+        super("Operator [" + operator + "] doesn't have a file in the repo");
+    }
+}
